♻️ refactor: refactor the chat store to slice mode
