--- src/store/chat/selectors/index.ts
@@ -1,3 +1,0 @@
-export { chatSelectors } from './chat';
-export { chatEnhanceSelectors } from './enchance';
-export { topicSelectors } from './topic';
